db/QYOG-176: migration pk키 전체 bigint format으로 변경 및 현재 소스코드 내의 엔티티 메타데이터와 migration 내의 코드 충돌로 인한 에러 수정
https://dongurami.atlassian.net/browse/QYOG-176

+++ migrations/club/1712482591735-club-member.ts
@@ -14,14 +14,14 @@ export class ClubMember1712482591735 implements MigrationInterface {
           generatePrimaryColumn('동아리 구성원'),
           {
             name: 'user_id',
-            type: 'int',
+            type: 'bigint',
             unsigned: true,
             isNullable: false,
             comment: '동아리 구성원 유저 고유 ID',
           },
           {
             name: 'club_id',
-            type: 'int',
+            type: 'bigint',
             unsigned: true,
             isNullable: false,
             comment: '동아리 고유 ID',
