+++ migrations/club/1713083761184-club-application.ts
@@ -17,14 +17,14 @@ export class ClubApplication1713083761184 implements MigrationInterface {
           generatePrimaryColumn('동아리 지원서'),
           {
             name: 'club_id',
-            type: 'int',
+            type: 'bigint',
             unsigned: true,
             isNullable: false,
             comment: '동아리 고유 ID',
           },
           {
             name: 'user_id',
-            type: 'int',
+            type: 'bigint',
             unsigned: true,
             isNullable: false,
             comment: '지원 유저 고유 ID',
